move editor redux
